Prevent // between logical_path and "index" but using File.join.
PATH_INFO is sometimes "/" instead of "" (I think this depends
on your web server and the Rack spec allows both). But the
previous code of logical_path + "/index" assumes the final "/"
was missing. Since this was not the case the end result was
//index.html which Hike could not find (it can find just
/index.html). So changed code to use File.join. This allows
the "/" in PATH_INFO to be there or not be there.