Add block to Brochure.app.
This allows additional manipulation to be done to the app before it is wrapped
in other middleware (Static, Failsafe, etc.). This is necessary because the
configuration options will never provide every option needed. This hook allows
a developer to bypass using supported options and just manipulate the app
directly.

The example provided by the tests is adding custom new template paths
to the template_trail outside of the standard "templates" directory.